[fix] refactoring
