+++ DdokBaro/Model/Model.swift
@@ -1,0 +1,12 @@
+//
+//  Model.swift
+//  DdokBaro
+//
+//  Created by yusang on 2023/08/27.
+//
+
+import Foundation
+
+struct airPod{
+    var degree:Int
+}
